--- dev.sh
@@ -1,3 +1,0 @@
-#!/bin/sh
-
-cd web && yarn dev & air && fg
