+++ testing/endtoend/backend/anvil_local_test.go
@@ -17,17 +17,8 @@ func TestLocalAnvilLoadAccounts(t *testing.T) {
 	if err := a.loadAccounts(); err != nil {
 		t.Fatal(err)
 	}
-	if a.alice == nil {
-		t.Error("Alice is nil")
-	}
-	if a.bob == nil {
-		t.Error("Bob is nil")
-	}
-	if a.charlie == nil {
-		t.Error("Charlie is nil")
-	}
-	if a.deployer == nil {
-		t.Error("Deployer is nil")
+	if len(a.accounts) == 0 {
+		t.Error("No accounts generated")
 	}
 }
 
@@ -38,23 +29,19 @@ func TestLocalAnvilStarts(t *testing.T) {
 	if err != nil {
 		t.Fatal(err)
 	}
-	if err := a.Start(); err != nil {
+	if err := a.Start(ctx); err != nil {
 		t.Fatal(err)
 	}
-	if _, err := a.DeployRollup(); err != nil {
+	if _, err := a.DeployRollup(ctx); err != nil {
 		t.Fatal(err)
 	}
 
 	// There should be at least 100 blocks
-	bn, err2 := a.Client().BlockNumber(ctx)
+	bn, err2 := a.Client().HeaderByNumber(ctx, nil)
 	if err2 != nil {
 		t.Fatal(err2)
 	}
-	if bn < 100 {
+	if bn.Number.Uint64() < 100 {
 		t.Errorf("Expected at least 100 blocks at start, but got only %d", bn)
-	}
-
-	if err := a.Stop(); err != nil {
-		t.Fatal(err)
 	}
 }
