+++ testing/endtoend/backend/backend.go
@@ -4,34 +4,25 @@
 package backend
 
 import (
+	"context"
+
 	challenge_testing "github.com/OffchainLabs/bold/testing"
 	"github.com/OffchainLabs/bold/testing/setup"
 	"github.com/ethereum/go-ethereum/accounts/abi/bind"
 	"github.com/ethereum/go-ethereum/common"
-	"github.com/ethereum/go-ethereum/ethclient"
 )
 
 type Backend interface {
-	// Operations
 	// Start sets up the backend and waits until the process is in a ready state.
-	Start() error
-	// Stop tears down the backend.
-	Stop() error
+	Start(ctx context.Context) error
 	// Client returns the backend's client.
-	Client() *ethclient.Client
-
-	// Transactors
-	// Alice represents the transactor for Alice's account.
-	Alice() *bind.TransactOpts
-	// Bob represents the transactor for Bob's account.
-	Bob() *bind.TransactOpts
-	// Charlie represents the transactor for Charlie's account.
-	Charlie() *bind.TransactOpts
-
-	// Deployer functions
+	Client() setup.Backend
+	// Accounts managed by the backend.
+	Accounts() []*bind.TransactOpts
 	// DeployRollup contract, if not already deployed.
-	DeployRollup(opts ...challenge_testing.Opt) (common.Address, error)
-
+	DeployRollup(ctx context.Context, opts ...challenge_testing.Opt) (common.Address, error)
 	// Contract addresses relevant to the challenge protocol.
 	ContractAddresses() *setup.RollupAddresses
+	// Commit a tx to the backend, if possible (simulated backend requires this)
+	Commit() common.Hash
 }
