+++ testing/endtoend/flaky_eth_test.go
@@ -1,10 +1,13 @@
-//go:build falky_eth_test
-// +build falky_eth_test
-
 package endtoend
 
-import "testing"
+import "github.com/ethereum/go-ethereum/common"
 
-func TestChallengeProtocol_AliceAndBob_AnvilLocal_WithFlakyEthClient(t *testing.T) {
-	aliceAndBobInMiddleOfBlock(t, true)
+// import "testing"
+
+// func TestChallengeProtocol_AliceAndBob_AnvilLocal_WithFlakyEthClient(t *testing.T) {
+// 	aliceAndBobInMiddleOfBlock(t, true)
+// }
+
+func hi() common.Address {
+	return common.Address{}
 }
